+++ src/fidesops/task/graph_task.py
@@ -90,8 +90,8 @@ def retry(
                     raised_ex = ex
 
             self.log_end(action_type, raised_ex)
-            self.resources.request.cache_failed_step_and_collection(
-                failed_step=action_type, failed_collection=self.traversal_node.address
+            self.resources.request.cache_failed_collection_details(
+                step=action_type, collection=self.traversal_node.address
             )
             # Re-raise to stop privacy request execution on failure.
             raise raised_ex